requirements.txt
flask
python-pptx
pandas
openpyxl
xlsxwriter
sentence-transformers
torch
tk
aspose.slides
pywin32
